Add read-only/tool restrictions to all AI providers with comprehensive security documentation (#37)
* feat: Add read-only/tool restrictions to all AI providers

Implements comprehensive read-only restrictions across all AI providers
for enhanced security when using external AI models for code analysis.

## Changes by Provider

### Claude Provider
- Add ALLOWED_TOOLS and DISALLOWED_TOOLS lists with granular Bash patterns
- Add SHELL_COMMAND_WARNING for security awareness
- Update _build_command() to enforce tool restrictions via CLI flags
- Update security_flags: read_only=True

### Gemini Provider
- Add ALLOWED_TOOLS list with class/function name filtering
- Add PIPED_COMMAND_WARNING about piped command vulnerability
- Update _build_command() to enforce tool restrictions
- Update security_flags: read_only=True

### Opencode Provider
- Add READONLY_TOOLS_CONFIG with dual-layer protection
- Implement config file generation for permission enforcement
- Add SHELL_COMMAND_WARNING
- Update security_flags: read_only=True

### Cursor Agent Provider
- Add ALLOWED_TOOLS and DISALLOWED_TOOLS lists
- Implement temporary .cursor/cli-config.json generation
- Add permission system (Read/Write/Shell patterns)
- Add cleanup logic for temp config files
- Add SHELL_COMMAND_WARNING
- Update security_flags: read_only=True

### Codex Provider
- Add SANDBOX_ALLOWED_OPERATIONS and SANDBOX_BLOCKED_OPERATIONS (documentation)
- Add SANDBOX_WARNING explaining OS-level enforcement
- Update _build_prompt() to inject security warning
- Update security_flags: read_only=True
- Note: Uses native --sandbox read-only flag (most secure)

## Test Coverage
- Updated all provider tests to account for new restrictions
- Added specific tests for:
  - Config file generation and cleanup
  - Security warning injection
  - Tool restriction enforcement
- All tests passing (37 tests total across 5 providers)

## Security Model Comparison
- **Codex**: OS-level sandboxing (most robust) ✅
- **Claude**: Explicit allowlist/denylist via CLI flags
- **Gemini**: Tool name filtering with policy engine
- **Opencode**: Dual-layer (tool disable + permission denial)
- **Cursor Agent**: Config-based allowlist (weakest, documented)



* changes

* fixes

* docs: Update documentation for provider security feature

- Update CHANGELOG.md with v0.7.5 release notes
- Add installation guide, security docs, and provider comparisons
- Update README with architectural details and multi-provider support



---------

+++ src/claude_skills/claude_skills/common/ai_tools.py
@@ -697,8 +697,9 @@ def execute_tool(
     prompt: str,
     *,
     model: Optional[str] = None,
-    timeout: int = 90
+    timeout: float = 90
 ) -> ToolResponse:
+
     """
     Execute AI tool with a prompt and return structured response.
 
@@ -793,7 +794,7 @@ def execute_tools_parallel(
     prompt: str,
     *,
     models: Optional[dict[str, str]] = None,
-    timeout: int = 90
+    timeout: float = 90
 ) -> MultiToolResponse:
     """
     Execute multiple AI tools in parallel with same prompt.
